@@ -1,40 +1,35 @@
-package com.letgo.book.infrastructure.event.subscriber;
+package com.letgo.book.infrastructure.event.subscriber
 
-import com.letgo.book.application.changeTitle.ChangeTitleCommand;
-import com.letgo.book.domain.BookCreated;
-import com.letgo.shared.application.bus.command.CommandBus;
-import com.letgo.shared.application.event.DomainEventSubscriber;
-import com.letgo.shared.domain.DomainEvent;
-import com.letgo.shared.infrastructure.InfrastructureService;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
+import com.letgo.book.application.changeTitle.ChangeTitleCommand
+import com.letgo.book.domain.BookCreated
+import com.letgo.shared.application.bus.command.CommandBus
+import com.letgo.shared.application.event.DomainEventSubscriber
+import com.letgo.shared.domain.DomainEvent
+import com.letgo.shared.infrastructure.InfrastructureService
+import org.springframework.context.event.EventListener
+import org.springframework.scheduling.annotation.Async
+import java.util.*
 
 @Async
 @InfrastructureService
-final public class ShoutTitleOnBookCreated implements DomainEventSubscriber {
-    private final CommandBus commandBus;
-
-    public ShoutTitleOnBookCreated(CommandBus commandBus) {
-        this.commandBus = commandBus;
+class ShoutTitleOnBookCreated(
+    private val commandBus: CommandBus
+) : DomainEventSubscriber {
+    override fun isSubscribedTo(event: DomainEvent): Boolean {
+        return BookCreated::class.java == event.javaClass
     }
 
-    @Override
-    public boolean isSubscribedTo(DomainEvent event) {
-        return BookCreated.class.equals(event.getClass());
-    }
-
-    @Override
-    @EventListener(BookCreated.class)
-    public void consume(DomainEvent event) throws Throwable {
-        if (event instanceof BookCreated) {
-            BookCreated bookCreated = (BookCreated) event;
+    @EventListener(BookCreated::class)
+    @Throws(Throwable::class)
+    override fun consume(event: DomainEvent) {
+        if (event is BookCreated) {
             commandBus.dispatch(
-                    new ChangeTitleCommand(
-                            bookCreated.aggregateId(),
-                            bookCreated.title().toUpperCase(),
-                            bookCreated.titleCreatedAt().plusNanos(100000)
-                    )
-            );
+                ChangeTitleCommand(
+                    event.aggregateId(),
+                    event.title().uppercase(Locale.getDefault()),
+                    event.titleCreatedAt().plusNanos(100000)
+                )
+            )
         }
     }
 }
