@@ -1,36 +1,29 @@
-package com.letgo.book;
+package com.letgo.book
 
-import com.letgo.shared.application.bus.command.CommandHandler;
-import com.letgo.shared.application.bus.query.QueryHandler;
-import com.letgo.shared.domain.DomainService;
-import com.letgo.shared.infrastructure.InfrastructureService;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.FilterType;
+import com.letgo.shared.application.bus.command.CommandHandler
+import com.letgo.shared.application.bus.query.QueryHandler
+import com.letgo.shared.domain.DomainService
+import com.letgo.shared.infrastructure.InfrastructureService
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.FilterType
 
 @SpringBootApplication
 @ComponentScan(
-        includeFilters = {
-                @ComponentScan.Filter(
-                        type = FilterType.ASSIGNABLE_TYPE,
-                        classes = {
-                                QueryHandler.class,
-                                CommandHandler.class,
-                        }
-                ),
-                @ComponentScan.Filter(
-                        type = FilterType.ANNOTATION,
-                        classes = {
-                                DomainService.class,
-                                InfrastructureService.class,
-                        }
-                ),
-        },
-        basePackages = {"com.letgo.book", "com.letgo.shared"}
+    includeFilters = [ComponentScan.Filter(
+        type = FilterType.ASSIGNABLE_TYPE,
+        classes = [QueryHandler::class, CommandHandler::class]
+    ), ComponentScan.Filter(
+        type = FilterType.ANNOTATION,
+        classes = [DomainService::class, InfrastructureService::class]
+    )], basePackages = ["com.letgo.book", "com.letgo.shared"]
 )
-public class App {
-    public static void main(String... args) {
-        SpringApplication.run(App.class, args);
+open class App {
+    companion object {
+        @JvmStatic
+        fun main(args: Array<String>) {
+            SpringApplication.run(App::class.java, *args)
+        }
     }
 }
