@@ -1,15 +1,11 @@
-package com.letgo.book.application.find;
+package com.letgo.book.application.find
 
-import com.letgo.shared.application.bus.query.Query;
+import com.letgo.shared.application.bus.query.Query
 
-final public class FindBookQuery implements Query {
-    private final String id;
-
-    public FindBookQuery(String id) {
-        this.id = id;
-    }
-
-    public String id() {
-        return id;
+class FindBookQuery(
+    private val id: String
+) : Query {
+    fun id(): String {
+        return id
     }
 }
