@@ -1,34 +1,26 @@
-package com.letgo.book.domain;
+package com.letgo.book.domain
 
-import com.letgo.shared.domain.DomainEvent;
+import com.letgo.shared.domain.DomainEvent
+import java.time.LocalDateTime
 
-import java.time.LocalDateTime;
-
-final public class BookCreated extends DomainEvent {
-    private final String title;
-    private final LocalDateTime titleCreatedAt;
-
-    public BookCreated(String id, String title, LocalDateTime titleCreatedAt) {
-        super(id);
-        this.title = title;
-        this.titleCreatedAt = titleCreatedAt;
+class BookCreated(
+    id: String,
+    private val title: String,
+    private val titleCreatedAt: LocalDateTime
+) : DomainEvent(id) {
+    fun title(): String {
+        return title
     }
 
-    public String title() {
-        return title;
+    fun titleCreatedAt(): LocalDateTime {
+        return titleCreatedAt
     }
 
-    public LocalDateTime titleCreatedAt() {
-        return titleCreatedAt;
+    override fun name(): String {
+        return "BookCreated"
     }
 
-    @Override
-    public String name() {
-        return "BookCreated";
-    }
-
-    @Override
-    public String body() {
-        return "{'title': " + title + ", 'titleCreatedAt': " + titleCreatedAt.toString() + "}";
+    override fun body(): String {
+        return "{'title': $title, 'titleCreatedAt': $titleCreatedAt}"
     }
 }
