@@ -1,11 +1,9 @@
-package com.letgo.book.domain;
+package com.letgo.book.domain
 
-final public class BookNotFound extends IndexOutOfBoundsException {
-    private BookNotFound(String message) {
-        super(message);
-    }
-
-    public static BookNotFound withId(BookId id) {
-        return new BookNotFound("Book not found with id " + id.value());
+class BookNotFound private constructor(message: String) : IndexOutOfBoundsException(message) {
+    companion object {
+        fun withId(id: BookId): BookNotFound {
+            return BookNotFound("Book not found with id " + id.value())
+        }
     }
 }
