@@ -1,18 +1,17 @@
-package com.letgo.book.infrastructure.configuration;
+package com.letgo.book.infrastructure.configuration
 
-import com.letgo.book.infrastructure.interceptor.CustomInterceptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import com.letgo.book.infrastructure.interceptor.CustomInterceptor
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @Configuration
-public class WebConfiguration implements WebMvcConfigurer {
+open class WebConfiguration(
     @Autowired
-    private CustomInterceptor customInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(customInterceptor).addPathPatterns("/book/**");
+    private val customInterceptor: CustomInterceptor
+) : WebMvcConfigurer {
+    override fun addInterceptors(registry: InterceptorRegistry) {
+        registry.addInterceptor(customInterceptor).addPathPatterns("/book/**")
     }
 }
