@@ -1,7 +1,6 @@
-package com.letgo.book.infrastructure.controller;
+package com.letgo.book.infrastructure.controller
 
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMapping
 
 @RequestMapping("/book")
-public abstract class BookController {
-}
+abstract class BookController
