@@ -1,19 +1,16 @@
-package com.letgo.book.infrastructure.persistence.mapping;
+package com.letgo.book.infrastructure.persistence.mapping
 
-import com.letgo.book.domain.BookTitle;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import com.letgo.book.domain.BookTitle
+import javax.persistence.AttributeConverter
+import javax.persistence.Converter
 
 @Converter
-final public class BookTitleConverter implements AttributeConverter<BookTitle, String> {
-    @Override
-    public String convertToDatabaseColumn(BookTitle attribute) {
-        return attribute == null ? null : attribute.value();
+class BookTitleConverter : AttributeConverter<BookTitle, String> {
+    override fun convertToDatabaseColumn(attribute: BookTitle): String {
+        return attribute.value()
     }
 
-    @Override
-    public BookTitle convertToEntityAttribute(String dbData) {
-        return dbData == null ? null : BookTitle.create(dbData);
+    override fun convertToEntityAttribute(dbData: String?): BookTitle? {
+        return if (dbData == null) null else BookTitle.create(dbData)
     }
 }
