@@ -1,5 +1,6 @@
-package com.letgo.shared.application.bus.command;
+package com.letgo.shared.application.bus.command
 
-public interface CommandBus {
-    public void dispatch(Command command) throws Exception;
+interface CommandBus {
+    @Throws(Exception::class)
+    fun dispatch(command: Command)
 }
