@@ -1,9 +1,7 @@
-package com.letgo.shared.application.event;
+package com.letgo.shared.application.event
 
-import com.letgo.shared.domain.DomainEvent;
+import com.letgo.shared.domain.DomainEvent
 
-import java.util.List;
-
-public interface DomainEventPublisher {
-    public void publish(List<DomainEvent> events);
+interface DomainEventPublisher {
+    fun publish(events: List<DomainEvent>)
 }
