@@ -1,24 +1,18 @@
-package com.letgo.shared.infrastructure.event.publisher;
+package com.letgo.shared.infrastructure.event.publisher
 
-import com.letgo.shared.application.event.DomainEventPublisher;
-import com.letgo.shared.domain.DomainEvent;
-import org.springframework.context.ApplicationEventPublisher;
+import com.letgo.shared.application.event.DomainEventPublisher
+import com.letgo.shared.domain.DomainEvent
+import org.springframework.context.ApplicationEventPublisher
+import java.util.function.Consumer
 
-import java.util.List;
-
-final public class SpringSyncDomainEventPublisher implements DomainEventPublisher {
-    private final ApplicationEventPublisher publisher;
-
-    public SpringSyncDomainEventPublisher(ApplicationEventPublisher publisher) {
-        this.publisher = publisher;
-    }
-
-    @Override
-    public void publish(List<DomainEvent> events) {
+class SpringSyncDomainEventPublisher(
+    private val publisher: ApplicationEventPublisher
+) : DomainEventPublisher {
+    override fun publish(events: List<DomainEvent>) {
         try {
-            events.forEach(publisher::publishEvent);
-        } catch (Throwable throwable) {
-            throwable.printStackTrace();
+            events.forEach(Consumer { event: DomainEvent -> publisher.publishEvent(event) })
+        } catch (throwable: Throwable) {
+            throwable.printStackTrace()
         }
     }
 }
